test(graphql): split free-text worker coverage
