+++ packages/graphql/test/freeTextEvaluation.worker.test.ts
@@ -15,14 +15,8 @@ import {
   createFreeTextAttempt,
   decideSemanticEvaluationConsent,
   getFreeTextPracticeState,
-  markFreeTextAttemptUnavailable,
-  retryFreeTextEvaluation,
-  revealFreeTextSolution,
 } from '../src/services/freeTextEvaluation.js'
-import {
-  handleEvaluateFreeTextAttempt,
-  handleEvaluateFreeTextAttemptFailure,
-} from '../src/services/freeTextEvaluationHandler.js'
+import { handleEvaluateFreeTextAttempt } from '../src/services/freeTextEvaluationHandler.js'
 import { applyEvaluatedFreeTextAttempt } from '../src/services/freeTextPracticeResponseApplication.js'
 import {
   cleanupFixtures,
@@ -30,10 +24,9 @@ import {
   evaluatorResponse,
   participantContext,
   semanticConfig,
-  workflowRunRef,
 } from './freeTextEvaluation.fixture.js'
 
-const TEST_PREFIX = `free-text-evaluation-worker-${Date.now()}`
+const TEST_PREFIX = `free-text-evaluation-worker-application-${Date.now()}`
 type Fixture = Awaited<ReturnType<typeof createFixture>>
 let fixture: Fixture
 
@@ -553,463 +546,5 @@ describe('semantic free-text evaluation worker', () => {
     expect(storedAttempt.evaluationStatus).toBe('PENDING')
     expect(storedAttempt.cycle.status).toBe('ACTIVE')
     expect(storedAttempt.questionResponseDetailId).toBeNull()
-  })
-
-  it.each([
-    'retry',
-    'reveal',
-  ] as const)('does not let stale worker completion overwrite a %s transition', async (transition) => {
-    const schedule = vi.fn().mockResolvedValue(workflowRunRef())
-    const ctx = participantContext(fixture.participant.id, schedule)
-    await decideSemanticEvaluationConsent(
-      { disclosureVersion: '2026-08-18', accepted: true },
-      ctx
-    )
-    const pending = await createFreeTextAttempt(
-      {
-        instanceId: fixture.instance.id,
-        answer: 'It spreads investments across assets.',
-        answerTime: 3,
-        clientSubmissionId: randomUUID(),
-      },
-      ctx,
-      { disclosureVersion: '2026-08-18' }
-    )
-    const attempt = pending.currentAttempt!
-    await markFreeTextAttemptUnavailable(
-      {
-        attemptId: attempt.id,
-        evaluationRevision: attempt.evaluationRevision,
-        reason: 'EVALUATOR_RESULT_UNAVAILABLE',
-        retryable: true,
-      },
-      prisma
-    )
-    const participantTransition =
-      transition === 'retry'
-        ? retryFreeTextEvaluation({ attemptId: attempt.id }, ctx, {
-            disclosureVersion: '2026-08-18',
-          })
-        : revealFreeTextSolution({ cycleId: pending.cycleId }, ctx)
-    const state = await participantTransition
-    const workerApplied = await prisma.$transaction((tx) =>
-      completeFreeTextAttemptEvaluationInTransaction(
-        {
-          attemptId: attempt.id,
-          evaluationRevision: attempt.evaluationRevision,
-          evaluation: evaluatorResponse(attempt.id, 60, 'partial'),
-        },
-        tx
-      )
-    )
-    expect(workerApplied).toBe(false)
-    expect(state.currentAttempt?.evaluationSource).toBeNull()
-    if (transition === 'retry') {
-      expect(state).toMatchObject({
-        cycleStatus: 'ACTIVE',
-        currentAttempt: {
-          evaluationRevision: 1,
-          evaluationStatus: 'PENDING',
-        },
-      })
-    } else {
-      expect(state).toMatchObject({
-        cycleStatus: 'SOLUTION_REVEALED',
-        currentAttempt: {
-          evaluationRevision: 0,
-          evaluationStatus: 'UNAVAILABLE',
-        },
-      })
-    }
-  })
-
-  it('keeps an uncertain non-match unavailable', async () => {
-    const ctx = participantContext(fixture.participant.id)
-    await decideSemanticEvaluationConsent(
-      { disclosureVersion: '2026-08-18', accepted: true },
-      ctx
-    )
-    const pending = await createFreeTextAttempt(
-      {
-        instanceId: fixture.instance.id,
-        answer: 'It spreads investments across assets.',
-        answerTime: 3,
-        clientSubmissionId: randomUUID(),
-      },
-      ctx,
-      { disclosureVersion: '2026-08-18' }
-    )
-    const response = evaluatorResponse(
-      pending.currentAttempt!.id,
-      60,
-      'partial'
-    )
-    response.rubric_assessments[0]!.needs_review = true
-    vi.stubGlobal(
-      'fetch',
-      vi.fn().mockResolvedValue(
-        new Response(JSON.stringify(response), {
-          status: 200,
-          headers: { 'content-type': 'application/json' },
-        })
-      )
-    )
-
-    await handleEvaluateFreeTextAttempt(
-      {
-        attemptId: pending.currentAttempt!.id,
-        evaluationRevision: 0,
-      },
-      { prisma } as never,
-      {} as never
-    )
-    const state = await getFreeTextPracticeState(
-      { instanceId: fixture.instance.id },
-      ctx
-    )
-    expect(state).toMatchObject({
-      attemptsUsed: 1,
-      attemptsRemaining: 1,
-      canSubmitAnswer: true,
-      currentAttempt: {
-        evaluationStatus: 'UNAVAILABLE',
-        evaluationSource: null,
-        correctness: null,
-        availabilityReason: 'EVALUATOR_RESULT_UNAVAILABLE',
-        retryable: false,
-      },
-    })
-  })
-
-  it('uses exact matching when an evaluator result is uncertain', async () => {
-    const ctx = participantContext(fixture.participant.id)
-    await decideSemanticEvaluationConsent(
-      { disclosureVersion: '2026-08-18', accepted: true },
-      ctx
-    )
-    const pending = await createFreeTextAttempt(
-      {
-        instanceId: fixture.instance.id,
-        answer: semanticConfig.accepted_exact_answers[0]!,
-        answerTime: 3,
-        clientSubmissionId: randomUUID(),
-      },
-      ctx,
-      { disclosureVersion: '2026-08-18' }
-    )
-    const response = evaluatorResponse(
-      pending.currentAttempt!.id,
-      60,
-      'partial'
-    )
-    response.rubric_assessments[0]!.needs_review = true
-    vi.stubGlobal(
-      'fetch',
-      vi.fn().mockResolvedValue(
-        new Response(JSON.stringify(response), {
-          status: 200,
-          headers: { 'content-type': 'application/json' },
-        })
-      )
-    )
-
-    await handleEvaluateFreeTextAttempt(
-      {
-        attemptId: pending.currentAttempt!.id,
-        evaluationRevision: 0,
-      },
-      { prisma } as never,
-      {} as never
-    )
-
-    const storedAttempt = await prisma.freeTextAttempt.findUniqueOrThrow({
-      where: { id: pending.currentAttempt!.id },
-      include: { cycle: true },
-    })
-    expect(storedAttempt).toMatchObject({
-      evaluationStatus: 'EVALUATED',
-      evaluationSource: 'EXACT_MATCH',
-      aggregateScore: 100,
-      correctness: 'CORRECT',
-      availabilityReason: 'EVALUATOR_RESULT_UNAVAILABLE',
-      retryable: false,
-      cycle: {
-        status: 'CORRECT',
-        bestScore: 100,
-        pointsAwarded: 10,
-      },
-    })
-    expect(storedAttempt.questionResponseDetailId).not.toBeNull()
-  })
-
-  it('uses exact matching after evaluator retries are exhausted exactly once', async () => {
-    const ctx = participantContext(fixture.participant.id)
-    await decideSemanticEvaluationConsent(
-      { disclosureVersion: '2026-08-18', accepted: true },
-      ctx
-    )
-    const pending = await createFreeTextAttempt(
-      {
-        instanceId: fixture.instance.id,
-        answer: semanticConfig.accepted_exact_answers[0]!,
-        answerTime: 3,
-        clientSubmissionId: randomUUID(),
-      },
-      ctx,
-      { disclosureVersion: '2026-08-18' }
-    )
-    const input = {
-      attemptId: pending.currentAttempt!.id,
-      evaluationRevision: 0,
-    }
-
-    const first = await handleEvaluateFreeTextAttemptFailure(
-      input,
-      { prisma } as never,
-      {} as never
-    )
-    const duplicate = await handleEvaluateFreeTextAttemptFailure(
-      input,
-      { prisma } as never,
-      {} as never
-    )
-
-    expect(first).toEqual({ success: true, applied: true })
-    expect(duplicate).toEqual({ success: true, applied: false })
-    expect(
-      await prisma.freeTextAttempt.findUniqueOrThrow({
-        where: { id: input.attemptId },
-        select: {
-          evaluationStatus: true,
-          evaluationSource: true,
-          aggregateScore: true,
-          questionResponseDetailId: true,
-        },
-      })
-    ).toMatchObject({
-      evaluationStatus: 'EVALUATED',
-      evaluationSource: 'EXACT_MATCH',
-      aggregateScore: 100,
-    })
-    expect(
-      await prisma.questionResponseDetail.count({
-        where: { freeTextAttempt: { id: input.attemptId } },
-      })
-    ).toBe(1)
-  })
-
-  it('uses exact matching when semantic evaluation cannot be scheduled', async () => {
-    const schedule = vi.fn().mockRejectedValue(new Error('scheduler offline'))
-    const ctx = participantContext(fixture.participant.id, schedule)
-    const consoleError = vi.spyOn(console, 'error').mockImplementation(() => {})
-    await decideSemanticEvaluationConsent(
-      { disclosureVersion: '2026-08-18', accepted: true },
-      ctx
-    )
-
-    const state = await createFreeTextAttempt(
-      {
-        instanceId: fixture.instance.id,
-        answer: semanticConfig.accepted_exact_answers[0]!,
-        answerTime: 3,
-        clientSubmissionId: randomUUID(),
-      },
-      ctx,
-      { disclosureVersion: '2026-08-18' }
-    )
-
-    expect(state).toMatchObject({
-      cycleStatus: 'CORRECT',
-      attemptsUsed: 1,
-      currentAttempt: {
-        evaluationStatus: 'EVALUATED',
-        evaluationSource: 'EXACT_MATCH',
-        correctness: 'CORRECT',
-      },
-    })
-    expect(consoleError).toHaveBeenCalledOnce()
-  })
-
-  it('ends the cycle after the configured evaluated attempts and reveals the solution', async () => {
-    const ctx = participantContext(fixture.participant.id)
-    await decideSemanticEvaluationConsent(
-      { disclosureVersion: '2026-08-18', accepted: true },
-      ctx
-    )
-
-    for (const score of [0, 0]) {
-      const pending = await createFreeTextAttempt(
-        {
-          instanceId: fixture.instance.id,
-          answer: `Incomplete answer ${score}`,
-          answerTime: 2,
-          clientSubmissionId: randomUUID(),
-        },
-        ctx,
-        { disclosureVersion: '2026-08-18' }
-      )
-      vi.stubGlobal(
-        'fetch',
-        vi
-          .fn()
-          .mockResolvedValue(
-            new Response(
-              JSON.stringify(
-                evaluatorResponse(pending.currentAttempt!.id, score, 'missing')
-              ),
-              { status: 200, headers: { 'content-type': 'application/json' } }
-            )
-          )
-      )
-      await handleEvaluateFreeTextAttempt(
-        {
-          attemptId: pending.currentAttempt!.id,
-          evaluationRevision: 0,
-        },
-        { prisma } as never,
-        {} as never
-      )
-    }
-
-    const exhausted = await getFreeTextPracticeState(
-      { instanceId: fixture.instance.id },
-      ctx
-    )
-    expect(exhausted).toMatchObject({
-      cycleStatus: 'EXHAUSTED',
-      attemptsUsed: 2,
-      attemptsRemaining: 0,
-      solutionAuthorized: true,
-      referenceSolution: semanticConfig.reference_solution,
-      canPracticeAgain: true,
-      canSubmitAnswer: false,
-    })
-    await expect(
-      createFreeTextAttempt(
-        {
-          instanceId: fixture.instance.id,
-          answer: 'A third answer',
-          answerTime: 2,
-          clientSubmissionId: randomUUID(),
-        },
-        ctx
-      )
-    ).rejects.toThrow('Start a new free-text practice cycle')
-  })
-
-  it('rechecks lecturer entitlement before external evaluation', async () => {
-    const ctx = participantContext(fixture.participant.id)
-    await decideSemanticEvaluationConsent(
-      { disclosureVersion: '2026-08-18', accepted: true },
-      ctx
-    )
-    const pending = await createFreeTextAttempt(
-      {
-        instanceId: fixture.instance.id,
-        answer: 'It spreads investments across assets.',
-        answerTime: 3,
-        clientSubmissionId: randomUUID(),
-      },
-      ctx,
-      { disclosureVersion: '2026-08-18' }
-    )
-    await prisma.user.update({
-      where: { id: fixture.lecturer.id },
-      data: { catalystIndividual: false, catalystInstitutional: false },
-    })
-    const fetchMock = vi.fn()
-    vi.stubGlobal('fetch', fetchMock)
-
-    await handleEvaluateFreeTextAttempt(
-      {
-        attemptId: pending.currentAttempt!.id,
-        evaluationRevision: 0,
-      },
-      { prisma } as never,
-      {} as never
-    )
-    const state = await getFreeTextPracticeState(
-      { instanceId: fixture.instance.id },
-      ctx
-    )
-    expect(state?.currentAttempt).toMatchObject({
-      evaluationStatus: 'UNAVAILABLE',
-      evaluationSource: null,
-      correctness: null,
-      availabilityReason: 'LECTURER_ENTITLEMENT_UNAVAILABLE',
-      retryable: true,
-    })
-    expect(state?.canSubmitAnswer).toBe(true)
-    expect(fetchMock).not.toHaveBeenCalled()
-  })
-
-  it('leaves an exhausted cycle terminal without solution details when reveal is disabled', async () => {
-    await prisma.elementInstance.update({
-      where: { id: fixture.instance.id },
-      data: {
-        elementData: {
-          ...fixture.instance.elementData,
-          options: {
-            ...(fixture.instance.elementData as { options: object }).options,
-            semanticEvaluation: {
-              ...semanticConfig,
-              solution_reveal_enabled: false,
-            },
-          },
-        } as never,
-      },
-    })
-    const ctx = participantContext(fixture.participant.id)
-    await decideSemanticEvaluationConsent(
-      { disclosureVersion: '2026-08-18', accepted: true },
-      ctx
-    )
-
-    for (const score of [0, 0]) {
-      const pending = await createFreeTextAttempt(
-        {
-          instanceId: fixture.instance.id,
-          answer: `Incomplete answer ${score}`,
-          answerTime: 2,
-          clientSubmissionId: randomUUID(),
-        },
-        ctx,
-        { disclosureVersion: '2026-08-18' }
-      )
-      vi.stubGlobal(
-        'fetch',
-        vi
-          .fn()
-          .mockResolvedValue(
-            new Response(
-              JSON.stringify(
-                evaluatorResponse(pending.currentAttempt!.id, score, 'missing')
-              ),
-              { status: 200, headers: { 'content-type': 'application/json' } }
-            )
-          )
-      )
-      await handleEvaluateFreeTextAttempt(
-        {
-          attemptId: pending.currentAttempt!.id,
-          evaluationRevision: pending.currentAttempt!.evaluationRevision,
-        },
-        { prisma } as never,
-        {} as never
-      )
-    }
-
-    const state = await getFreeTextPracticeState(
-      { instanceId: fixture.instance.id },
-      ctx
-    )
-    expect(state).toMatchObject({
-      cycleStatus: 'EXHAUSTED',
-      canPracticeAgain: true,
-      canRevealSolution: false,
-      solutionAuthorized: false,
-      referenceSolution: null,
-      explanation: null,
-    })
   })
 })
